Worked on review comments and fixed
