+++ src/Services/GetApi.js
@@ -1,0 +1,5 @@
+export class GetApi {
+    request(url, method, body) {
+      return fetch(url);
+    }
+  }
