+++ src/Services/PostApi.js
@@ -1,0 +1,7 @@
+class PostApi {
+    constructor(url) {
+    }
+    request(url, method, body) {
+      return fetch(url);
+    }
+  }
