+++ src/Services/PutApi.js
@@ -1,0 +1,7 @@
+class PutApi {
+    constructor(url) {
+    }
+    request(url, method, body) {
+      return fetch(url);
+    }
+  }
